some refactoring of comments
